acceso restringido en todas las paginas

+++ resources/templates/contenido/eliminarPublicacion.php
@@ -1,4 +1,10 @@
 <?php
+
+if($_SESSION['tipo_cliente'] == 'empresa'){
+  header('Location: accesoRestringido.php');
+  die();
+}
+
 $id_pub = $_GET['idPub'];
 
 PublicacionesManager::delete($id_pub);
